collector/requirements.txt
requests
sqlalchemy
psycopg2-binary
prettytable
Pillow
python-telegram-bot
httpx[socks]